Declare Redux (et. al.) as PeerDependencies
Redux and webpack-hot-middleware are currently required by the webpack config that builds storybook versions of your project's code.  It seems to be the case that the project webpack config needs to have access to these two modules.

I'm not sure if there's a way to explicitly require and expose them, but letting the developer know that they are required seems like a step in the right direction.